refactor: 정렬 버튼 공유 컴포넌트로 정의
