+++ app/src/main/java/com/example/linkit/presentation/view/Home/AppBarHome.kt
@@ -71,7 +71,7 @@ fun AppBarHome(navController: NavController) {
  *   - 탐색화면: 돋보기와 정렬 아이콘
  */
 @Composable
-fun AppBarHomeDefault(
+private fun AppBarHomeDefault(
     modifier: Modifier = Modifier,
     onSearchClicked: () -> Unit,
     onProfileClick: () -> Unit,
@@ -109,7 +109,7 @@ fun AppBarHomeDefault(
 
 @Preview
 @Composable
-fun PreviewAppBar() {
+private fun PreviewAppBar() {
     val navController = rememberNavController()
 
     LinkItTheme {
